CXBOX-207 statistic blocks

+++ src/main/java/org/demo/documentation/widgets/statsblock/base/data/MyEntity4201.java
@@ -15,5 +15,5 @@ public class MyEntity4201 extends BaseEntity {
     @Column
     private String customField;
     @Column
-    private Integer customFieldNum = 0;
+    private Long customFieldNum = 0L;
 }
