+++ src/main/java/org/demo/documentation/widgets/statsblock/base/data/MyExample4201DTO.java
@@ -10,7 +10,7 @@ import org.cxbox.api.data.dto.DataResponseDTO;
 @NoArgsConstructor
 public class MyExample4201DTO extends DataResponseDTO {
     private String customField;
-    private Integer customFieldNum;
+    private Long customFieldNum;
     public MyExample4201DTO(MyEntity4201 entity) {
         this.id = entity.getId().toString();
         this.customField = entity.getCustomField();
